+++ packages/manager/apps/pci-ai-tools/src/pages/qpus/create/_components/useOrderFunnel.hook.tsx
@@ -96,7 +96,7 @@ export function useOrderFunnel(
       )?.unsecureHttp
         ? PrivacyEnum.public
         : PrivacyEnum.private,
-      timeoutAutoRestart: false,
+      timeoutAutoRestart: true,
       labels: [],
       sshKey: [],
       volumes: [],