+++ packages/manager/apps/pci-instances/src/pages/instances/create/components/RegionSelectionModal.component.tsx
@@ -82,6 +82,7 @@ const RegionSelectionModal: FC<TRegionSelectionModalProps> = ({
         {children}
         <LocalizationSelect<TCustomRegionItemData>
           className="mt-4"
+          createPortal={false}
           regions={regions}
           onValueChange={handleSelect}
           value={selectRegionValue}