+++ packages/manager/apps/pci-instances/src/pages/instances/create/view-models/instanceNameViewModel.ts
@@ -1,0 +1,10 @@
+import { TInstancesCatalog } from '@/domain/entities/instancesCatalog';
+import { getOvhInstanceName } from '@/domain/services/instance.service';
+
+export const selectOvhInstanceName = (regionalizedFlavorId: string | null) => (
+  catalog?: TInstancesCatalog,
+) => {
+  if (!catalog || !regionalizedFlavorId) return null;
+
+  return getOvhInstanceName(catalog, regionalizedFlavorId);
+};