+++ packages/manager/apps/pci-instances/src/pages/instances/instance/Instance.page.tsx
@@ -69,7 +69,7 @@ const Instance: FC = () => {
               </div>
             </div>
           </header>
-          <div className="mt-8 mb-8">
+          <div className="my-8">
             <Notifications />
             <SearchNotifications />
           </div>