+++ packages/manager/apps/web-hosting/src/pages/dashboard/local-seo/__tests__/LocalSeo.page.spec.tsx
@@ -2,7 +2,7 @@ import { render, screen } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { wrapper } from '@/utils/test.provider';
-import { mockUseDataApi } from '@/utils/test.setup';
+import { getDomRect, mockUseDataApi } from '@/utils/test.setup';
 
 import LocalSeo from '../LocalSeo.page';
 
@@ -18,6 +18,7 @@ vi.mock('@/hooks/localSeo/useDatagridColumn', () => ({
 
 describe('LocalSeo page', () => {
   beforeEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(120, 120));
     vi.clearAllMocks();
     mockUseDataApi.mockReturnValue({
       flattenData: [],
@@ -34,6 +35,9 @@ describe('LocalSeo page', () => {
       totalCount: 0,
     });
   });
+  afterEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(0, 0));
+  });
 
   it('should render correctly', () => {
     const { container } = render(<LocalSeo />, { wrapper });
@@ -44,5 +48,11 @@ describe('LocalSeo page', () => {
     render(<LocalSeo />, { wrapper });
 
     expect(screen.getByRole('button')).toBeInTheDocument();
+  });
+  it('should have a valid html with a11y and w3c', async () => {
+    const { container } = render(<LocalSeo />, { wrapper });
+    const html = container.innerHTML;
+    await expect(html).toBeValidHtml();
+    await expect(container).toBeAccessible();
   });
 });