+++ packages/manager/apps/web-hosting/src/pages/dashboard/multisite/domain/__tests__/DetacheDomain.modal.spec.tsx
@@ -2,15 +2,18 @@ import { fireEvent, render } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { wrapper } from '@/utils/test.provider';
-import { navigate } from '@/utils/test.setup';
+import { getDomRect, navigate } from '@/utils/test.setup';
 
 import DetacheDomainModal from '../DetacheDomain.modal';
 
 describe('DetacheDomainModal', () => {
   beforeEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(120, 120));
     vi.clearAllMocks();
   });
-
+  afterEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(0, 0));
+  });
   it('should close modal on cancel', () => {
     const { getByTestId } = render(<DetacheDomainModal />, { wrapper });