+++ packages/manager/apps/web-hosting/src/pages/dashboard/multisite/git/__tests__/LastDeploymentGit.modal.spec.tsx
@@ -2,15 +2,18 @@ import { fireEvent, render } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { wrapper } from '@/utils/test.provider';
-import { navigate } from '@/utils/test.setup';
+import { getDomRect, navigate } from '@/utils/test.setup';
 
 import LastDeploymentGitModal from '../LastDeploymentGit.modal';
 
 describe('LastDeploymentGitModal', () => {
   beforeEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(120, 120));
     vi.clearAllMocks();
   });
-
+  afterEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(0, 0));
+  });
   it('should render the modal with the correct title', () => {
     const { getByTestId } = render(<LastDeploymentGitModal />, {
       wrapper,