+++ packages/manager/apps/web-hosting/src/pages/dashboard/multisite/module/__tests__/AddModule.modal.spec.tsx
@@ -2,14 +2,18 @@ import { render } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { wrapper } from '@/utils/test.provider';
+import { getDomRect } from '@/utils/test.setup';
 
 import AddModuleModal from '../AddModule.modal';
 
 describe('AddModuleModal', () => {
   beforeEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(120, 120));
     vi.clearAllMocks();
   });
-
+  afterEach(() => {
+    Element.prototype.getBoundingClientRect = vi.fn(() => getDomRect(0, 0));
+  });
   it('should render correctly', () => {
     const { container } = render(<AddModuleModal />, { wrapper });
     expect(container).toBeInTheDocument();