+++ packages/manager/apps/web-hosting/src/utils/tracking.constants.ts
@@ -100,5 +100,6 @@ export const REMOVE_SEO_SUBSCIPTION = 'remove-seo-subscription';
 
 export const WORDPRESS_MANAGED = 'managed-wordpress';
 export const WORDPRESS_MANAGED_SERVICE = 'managed-wordpress_service';
+export const WORDPRESS_MANAGED_ONBOARDING = 'managed-wordpress_onboarding';
 
 export const GO_TO = (link: string) => `go-to-${link}_${WEBSITE}`;