+++ packages/manager/modules/sms/CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.20.8](https://github.com/ovh/manager/compare/@ovh-ux/manager-sms@8.20.7...@ovh-ux/manager-sms@8.20.8) (2026-02-17)
+
+
+### Bug Fixes
+
+* **i18n:** add missing translations [CDS 4663] ([3c441dc](https://github.com/ovh/manager/commit/3c441dc5b7f9a9af450bd12cd16385c882e1112b))
+* **i18n:** add missing translations [CDS 4668] ([cc5b969](https://github.com/ovh/manager/commit/cc5b969630c8a964fda88c38a8e19c7d7fe41af6))
+* **i18n:** add missing translations [CDS 4669] ([81247b5](https://github.com/ovh/manager/commit/81247b5f4510e1dfa0b251ecdedb938504b3a94f))
+
+
+
+
+
 ## [8.20.7](https://github.com/ovh/manager/compare/@ovh-ux/manager-sms@8.20.6...@ovh-ux/manager-sms@8.20.7) (2026-02-16)